Moved methods from SplitStatevectorSimulatorPy to QasmSimulatorPy where they belong
