+++ qiskit/providers/basicaer/split_statevector_simulator.py
@@ -30,13 +30,9 @@ for the different splits, kept under statevector_tree
 """
 
 import logging
-import copy
 from math import log2
-import numpy as np
 from qiskit.util import local_hardware_info
-from qiskit.providers.basicaer.exceptions import BasicAerError
 from qiskit.providers.models import QasmBackendConfiguration
-from .qasm_simulator import QasmSimulatorPy
 from .statevector_simulator import StatevectorSimulatorPy
 
 logger = logging.getLogger(__name__)
@@ -103,9 +99,6 @@ class SplitStatevectorSimulatorPy(StatevectorSimulatorPy):
         super().__init__(configuration=(
             configuration or QasmBackendConfiguration.from_dict(self.DEFAULT_CONFIGURATION)),
                          provider=provider)
-        # _substates field will contain 2 statevectors corresponding to measurement results 0 and 1
-        self._substates = []
-        self._substate_probabilities = []
 
     def run(self, qobj, backend_options=None):
         """Run qobj asynchronously.
@@ -140,102 +133,5 @@ class SplitStatevectorSimulatorPy(StatevectorSimulatorPy):
                 }
         """
         return super().run(qobj, backend_options=backend_options)
-
-    def _get_single_outcome(self, qubit, wanted_state):
-        """Simulate the outcome of a determined measurement of a qubit.
-        This function mainly exists to perserve the existing function division,
-        replacing _get_measure_outcome which is called for the other simulators
-
-        Args:
-            qubit (int): the qubit to measure
-            wanted_state (int): the outcome state to return
-
-        Return:
-            tuple: pair (outcome, probability) where outcome is '0' or '1' and
-            probability is the probability of the returned outcome.
-        """
-        # Axis for numpy.sum to compute probabilities
-        axis = list(range(self._number_of_qubits))
-        axis.remove(self._number_of_qubits - 1 - qubit)
-        probabilities = np.sum(np.abs(self._statevector) ** 2, axis=tuple(axis))
-        # Compute einsum index string for 1-qubit matrix multiplication
-
-        return str(wanted_state), probabilities[wanted_state]
-
-    def _split_statevector(self, qubit, probabilities, cmembit, cregbit=None):
-        """Split the statevector into two substates corresponding to the two measurement options.
-
-        Args:
-            qubit (int): the measured qubit
-            probabilities (list[float]): the probabilities of measuring 0 or 1
-            cmembit (int): the classical memory bit to store outcome in.
-            cregbit (int, optional): is the classical register bit to store outcome in.
-
-        Raises:
-            BasicAerError: if an error occurred.
-        """
-        # Getting the data for each split possibility
-        outcome0, probability0 = self._get_single_outcome(qubit, 0)
-        outcome1, probability1 = self._get_single_outcome(qubit, 1)
-
-        # Copying the statevector into its two substates
-        temp = copy.deepcopy(self)
-        self._substates.append(copy.deepcopy(temp))
-        self._substate_probabilities.append(probabilities[0])
-        self._substates.append(copy.deepcopy(temp))
-        self._substate_probabilities.append(probabilities[1])
-
-        # Updating the states after the split
-        self._substates[0]._update_state_after_measure(qubit, cmembit,
-                                                       outcome0, probability0, cregbit)
-        self._substates[1]._update_state_after_measure(qubit, cmembit,
-                                                       outcome1, probability1, cregbit)
-
-    def _generate_data(self):
-        """Run an experiment (circuit) and return a single experiment result.
-
-        Args:
-            experiment (QobjExperiment): experiment from qobj experiments list
-
-        Returns:
-             data: A result dictionary in the form of a binary tree which looks something like:
-
-                {
-                "value": The statevector of the state at the first measurement
-                "prob_0": The probability to get a measurement of 0 at the first measurement
-                "prob_1": The probability to get a measurement of 1 at the first measurement
-                "path_0":
-                    {
-                        "value": The statevector evolved from result 0 in the first measurement,
-                                 at the second measurement or at the end of the circuit
-                        "prob_0": The probability for a measurement of 0 at the second measurement
-                        "prob_1": The probability for a measurement of 1 at the second measurement
-                        "path_0":
-                            {
-                                ...
-                            }
-                        "path_1":
-                            {
-                                ...
-                            }
-                    }
-                "path_1":
-                    {
-                        ...
-                    }
-                }
-        Raises:
-            BasicAerError: if an error occurred.
-        """
-        if self._substates != []:
-            data = {'value': self._get_statevector(),
-                    'path_0': self._substates[0]._generate_data(),
-                    'path_0_probability': self._substate_probabilities[0],
-                    'path_1': self._substates[1]._generate_data(),
-                    'path_1_probability': self._substate_probabilities[1]}
-        else:
-            data = {'value': self._get_statevector()}
-
-        return data
 
     # TODO: Add a _validate method
